Fix cache name computation
This makes sure we are using the same algorithm to generate the cache name for
both remote and local cache.
